estimate-bot/requirements.txt
# Core dependencies
python-dotenv==1.0.0
ccxt==4.4.69
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
python-telegram-bot==20.7
Flask==3.0.0
Flask-SocketIO==5.3.6

# Data handling and analysis
pandas==2.1.4
numpy==1.26.2
ta-lib==0.4.28
python-binance==1.0.19
bybit-api==1.0.0

# Blockchain interaction
web3==6.11.3
xrpl-py==2.4.0
stellar-sdk==9.1.0
hedera-sdk-py==2.24.6
py-algorand-sdk==2.5.0

# Security and encryption
cryptography==41.0.7
pyjwt==2.8.0

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Development tools
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Install with: pip install -r requirements.txt 